+++ yesod-test/ChangeLog.md
@@ -1,3 +1,8 @@
+## 1.4.3.2
+
+* Add `addTokenFromCookie` and `addTokenFromCookieNamedToHeaderNamed`, which support the new CSRF token middleware [#1058](https://github.com/yesodweb/yesod/pull/1058)
+* Add `getRequestCookies`, which returns the cookies from the most recent request [#1058](https://github.com/yesodweb/yesod/pull/1058)
+
 ## 1.4.3.1
 
 * Improved README